Refactor: 펀딩 성공 시점을 마감일로 변경
